+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Synchronous Queue Support #11
+- Doctrine Queue Support #12
+- WrappedJob Helper methods
 
 ## 0.3.4 - 2017-06-03